EMP-148: Fixed error with returning query and amend to search by date parameters
EMP-148: Fixed error with returning query and amend to search by date parameters

+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/repository/CrmFormsViewRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import uk.gov.justice.laa.crime.equinity.historicaldata.model.CrmFormsViewModel;
 
 
-public interface CrmFormsViewRepository extends JpaRepository<CrmFormsViewModel, Long>, JpaSpecificationExecutor<CrmFormsViewModel> {
+public interface CrmFormsViewRepository extends JpaRepository<CrmFormsViewModel, Integer>, JpaSpecificationExecutor<CrmFormsViewModel> {
 }
